2_LLM_Demo_App/backend/requirements.txt
sentence-transformers==2.2.2
transformers==4.35.0
torch>=2.2.0
chromadb==0.4.0
fastapi>=0.68.0,<0.111.0
uvicorn==0.24.0
pydantic>=1.9,<2.0
numpy==1.26.0
Pillow==10.1.0
python-multipart==0.0.6
